% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxecat_over.R
\name{proxecat_over}
\alias{proxecat_over}
\title{Overdispersed ProxECAT}
\usage{
proxecat_over(x1, x2, x3, x4, size = 1000)
}
\arguments{
\item{x1}{vector of the number of functional rare alleles in cases for each gene region}

\item{x2}{vector of the number of synonymous rare alleles (or other proxy) in cases for each gene region}

\item{x3}{vector of the number of functional rare alleles in controls for each gene region}

\item{x4}{vector of the number of synonymous rare alleles (or other proxy) in controls for each gene region}

\item{size}{overdispersion parameter. The smaller size, the more over dispersed. The larger size the less overdispersed.  Size of >1000 is approximately Poisson.}
}
\value{
input data, likelihood ratio test statistic and p-value for overdispersed ProxECAT. LRT is asymptotically chi-squared with 1 df.
}
\description{
Runs the overdispersed Proxy External Controls Association Test (ProxECAT) method given the number of rare alleles in each gene region and an overdispersion parameter
}
\examples{
##runs overdispersed proxy-ECAT for 5 gene regions
x1<-c(12,10,5,103,89)
x2<-c(6,5,3,10,3)
x3<-c(20,20,8,5,8)
x4<-c(9,11,8,7,3)
temp<-proxecat_over(x1=x1, x2=x2, x3=x3, x4=x4, size=100)
temp

}
